+++ src/pmp/visualization/Shader.cpp
@@ -10,7 +10,7 @@
 
 namespace pmp {
 
-Shader::Shader() : pid_(0) {}
+Shader::Shader() = default;
 
 Shader::~Shader()
 {